+++ changelog.md
@@ -1,3 +1,5 @@
 # 14.5.0
 * Add support for v14 with DND5.x
-* 
+* Added Point Buy Calculator
+* Added logic to make sure the current combatant's token is always on top in the scene
+* Fixed all Drow NPC's that were missing their creature type